+++ include/DCF77Signal.h
