+++ include/JJYSignal.h
