+++ include/MSFSignal.h
